chore: re-factor blob storage
