+++ server/src/main/java/com/epam/aidial/core/server/controller/DownloadFileController.java
@@ -3,8 +3,8 @@ package com.epam.aidial.core.server.controller;
 import com.epam.aidial.core.server.Proxy;
 import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.resource.ResourceDescriptor;
-import com.epam.aidial.core.server.storage.InputStreamReader;
 import com.epam.aidial.core.server.util.HttpStatus;
+import com.epam.aidial.core.server.vertx.stream.InputStreamReader;
 import io.vertx.core.Future;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpServerResponse;
