+++ server/src/main/java/com/epam/aidial/core/server/controller/RouteController.java
@@ -8,9 +8,9 @@ import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.upstream.RouteEndpointProvider;
 import com.epam.aidial.core.server.upstream.UpstreamProvider;
 import com.epam.aidial.core.server.upstream.UpstreamRoute;
-import com.epam.aidial.core.server.util.BufferingReadStream;
 import com.epam.aidial.core.server.util.HttpStatus;
 import com.epam.aidial.core.server.util.ProxyUtil;
+import com.epam.aidial.core.server.vertx.stream.BufferingReadStream;
 import io.vertx.core.Future;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpClientRequest;
