+++ server/src/main/java/com/epam/aidial/core/server/controller/ShareController.java
@@ -15,11 +15,10 @@ import com.epam.aidial.core.server.security.EncryptionService;
 import com.epam.aidial.core.server.service.InvitationService;
 import com.epam.aidial.core.server.service.LockService;
 import com.epam.aidial.core.server.service.ShareService;
-import com.epam.aidial.core.server.storage.BlobStorageUtil;
+import com.epam.aidial.core.server.util.BucketBuilder;
 import com.epam.aidial.core.server.util.HttpException;
 import com.epam.aidial.core.server.util.HttpStatus;
 import com.epam.aidial.core.server.util.ProxyUtil;
-import com.epam.aidial.core.server.util.ResourceUtil;
 import io.vertx.core.Future;
 import io.vertx.core.buffer.Buffer;
 import lombok.extern.slf4j.Slf4j;
@@ -76,7 +75,7 @@ public class ShareController {
                         throw new IllegalArgumentException("Can't list shared resources. Incorrect body");
                     }
 
-                    String bucketLocation = BlobStorageUtil.buildInitiatorBucket(context);
+                    String bucketLocation = BucketBuilder.buildInitiatorBucket(context);
                     String bucket = encryptionService.encrypt(bucketLocation);
                     String with = request.getWith();
 
@@ -105,7 +104,7 @@ public class ShareController {
                         throw new IllegalArgumentException("Can't initiate share request. Incorrect body");
                     }
 
-                    String bucketLocation = BlobStorageUtil.buildInitiatorBucket(context);
+                    String bucketLocation = BucketBuilder.buildInitiatorBucket(context);
                     String bucket = encryptionService.encrypt(bucketLocation);
                     return proxy.getVertx().executeBlocking(() -> shareService.initializeShare(bucket, bucketLocation, request), false);
                 })
@@ -118,7 +117,7 @@ public class ShareController {
                 .body()
                 .compose(buffer -> {
                     ResourceLinkCollection request = getResourceLinkCollection(buffer, Operation.DISCARD);
-                    String bucketLocation = BlobStorageUtil.buildInitiatorBucket(context);
+                    String bucketLocation = BucketBuilder.buildInitiatorBucket(context);
                     String bucket = encryptionService.encrypt(bucketLocation);
                     return proxy.getVertx()
                             .executeBlocking(() -> {
@@ -135,11 +134,11 @@ public class ShareController {
                 .body()
                 .compose(buffer -> {
                     RevokeResourcesRequest request = getRevokeResourcesRequest(buffer, Operation.REVOKE);
-                    String bucketLocation = BlobStorageUtil.buildInitiatorBucket(context);
+                    String bucketLocation = BucketBuilder.buildInitiatorBucket(context);
                     String bucket = encryptionService.encrypt(bucketLocation);
                     Map<ResourceDescriptor, Set<ResourceAccessType>> permissionsToRevoke = request.getResources().stream()
                             .collect(Collectors.toUnmodifiableMap(
-                                    resource -> ResourceUtil.resourceFromUrl(resource.url(), encryptionService),
+                                    resource -> ShareService.resourceFromUrl(resource.url(), encryptionService),
                                     SharedResource::permissions));
                     return proxy.getVertx()
                             .executeBlocking(() -> lockService.underBucketLock(bucketLocation, () -> {
@@ -173,7 +172,7 @@ public class ShareController {
                         throw new IllegalArgumentException("destinationUrl must be provided");
                     }
 
-                    String bucketLocation = BlobStorageUtil.buildInitiatorBucket(context);
+                    String bucketLocation = BucketBuilder.buildInitiatorBucket(context);
                     String bucket = encryptionService.encrypt(bucketLocation);
 
                     ResourceDescriptor source = ResourceDescriptorFactory.fromPrivateUrl(sourceUrl, encryptionService);
